changed some variables and parameters type from long to size_t to remove build warnings for x64 platform

+++ DICOM [Native]/libijg16/jmemsys16.h
@@ -103,7 +103,7 @@ EXTERN(void) jpeg_free_large JPP((j_common_ptr cinfo, void FAR * object,
 EXTERN(long) jpeg_mem_available JPP((j_common_ptr cinfo,
 				     long min_bytes_needed,
 				     long max_bytes_needed,
-				     long already_allocated));
+				     size_t already_allocated));
 
 
 /*
